requirements.txt
mysql_connector_repackaged==0.3.1
rich==9.2.0
